+++ pages/2_🥃_DashBoard.py
@@ -1,0 +1,5 @@
+import streamlit as st 
+
+
+
+st.title("THIS IS THE CARLOS PAGE")
